Supports Scientific Equivalent
Fixes this >> http://prntscr.com/ijzbkj

for example:

The Scientific format displays a number in exponential notation, replacing part of the number with E+n, where E (which stands for Exponent) multiplies the preceding number by 10 to the nth power. For example, a 2-decimal Scientific format displays 12345678901 as 1.23E+10, which is 1.23 times 10 to the 10th power.