(!) instead of (not) fn_insideHouseConditions.sqf (#69)
* (!) instead of (not) fn_insideHouseConditions.sqf

* (for "_i") indefinite loop instead of )while {true}) loop :+1:

Better than *while* loop.

* Revert "(for "_i") indefinite loop instead of )while {true}) loop :+1:"

This reverts commit 31922ec79396a980b14bd5ad9581faf08d326604.

* Replaced (while loop) with (for "_i") :+1:

* Silly me >> scripted function with no suspension and I spawned it ;)

* [] spawn not needed! I'm a silly goose

[] Spawn not needed for scripted function with no suspension ....

* Supports Scientific Equivalent

Fixes this >> http://prntscr.com/ijzbkj

for example:

The Scientific format displays a number in exponential notation, replacing part of the number with E+n, where E (which stands for Exponent) multiplies the preceding number by 10 to the nth power. For example, a 2-decimal Scientific format displays 12345678901 as 1.23E+10, which is 1.23 times 10 to the 10th power.